Moved defs to common file
