+++ Tribler/Test/Core/test_session.py
@@ -11,8 +11,8 @@ from Tribler.Core.leveldbstore import LevelDbStore
 from Tribler.Core.simpledefs import NTFY_CHANNELCAST, DLSTATUS_STOPPED, SIGNAL_CHANNEL, SIGNAL_ON_CREATED
 from Tribler.Core.TorrentDef import TorrentDef
 from Tribler.Test.Core.base_test import TriblerCoreTest
+from Tribler.Test.common import TORRENT_FILE
 from Tribler.Test.test_as_server import TestAsServer
-from Tribler.Test.test_libtorrent_download import TORRENT_FILE
 from Tribler.dispersy.util import blocking_call_on_reactor_thread
